Merge conflicts test1.txt
When multiple developers try to change the same files at the same time, this can result in what is called a merge conflict. By default, git is pretty good at merging changes during pulls and pushes. Sometimes, however, it can't figure out the best choice, and so it leaves it up to you.